requirements.txt
streamlit>=1.25
opencv-python
Pillow
numpy
